more

+++ src/common.rs
@@ -2,7 +2,7 @@ pub const PRG_ROM_UNIT_SIZE: usize = 0x4000;
 pub const CHR_ROM_UNIT_SIZE: usize = 0x2000;
 pub const PRG_RAM_UNIT_SIZE: usize = 0x2000;
 
-pub const CPU_CYCLES_PER_FRAME: u16 = 29780;
+pub const CPU_CYCLES_PER_FRAME: usize = 29780;
 pub const FRAME_WIDTH: usize = 256;
 pub const FRAME_HEIGHT: usize = 240;
